Mark RequesterFactory as interface, introduce RequesterFactoryProxy as implementation
